Разбил функции на классы, добавил функциональности
