--- README.md
@@ -1,7 +1,0 @@
-# Work in progress
-
-Реализована тепловая карта с использованием функций
-
-# Todo:
-
-Разделить функции на методы соответствующих классов в отдельной ветке, слить ветки
